made tictactoe work by fixing bug with copy calls on while loop condition

+++ src/hfs/diagnostics/ir_lowerer_errors.rs
@@ -19,7 +19,7 @@ pub enum IrLowererErrorKind {
     StackUnderflow,
     ExpectedItemOnStack,
     ExpectedNetZeroStackEffectIfStmt(usize),
-    ExpectedNetZeroStackEffectWhileLoop(usize),
+    ExpectedNetZeroStackEffectWhileLoop(i64),
 }
 
 #[derive(Debug)]
